Add files via upload
AWS Project2 

Scenario:
In the connected world, it is imperative that the organizations be interlinked with the customers and vendors. This process has been very sluggish, manual, batch-based and prone to failures. Such integration design has lead to impaired decision-making and delays in the detection of fraudulent actions. This project created an automated, event-based real-time process using managed cloud services that do not have these limitations.

Skills and tools: AWS S3, AWS EC2, AWS SNS, AWS RDS, Python, Boto3
Architecture Implementation
The customer uploads the invoice data to S3 bucket in a text format as per their guidelines and policies. This bucket will have a policy to auto delete any content that is more than 1 day old (24 hours).
An event will trigger in the bucket that will place a message in SNS topic
A custom program running in EC2 will subscribe to the SNS topic and get the message placed by S3 event
The program will use S3 API to read from the bucket, parse the content of the file and create a CSV record and save the details in an RDS database
The program will use S3 API to write CSV record to destination S3 bucket as new S3 object.