Add files via upload
AWS Project 3			

Scenario:

Testing/development of a cloud-native application on the cloud might not be feasible due to cost constraints. Therefore, when working on an application using DynamoDB as the back-end, we have to deploy it on-prem first for debugging purposes. 

This project deploys DynamoDB using an IaaS deployment on EC2 instances. This easily be adapted to an on-premise installation for testing and debugging purposes.
Objectives
Install DynamoDB on an EC2 instance as an IaaS installation
Connect to the IaaS DynamoDB installation using the given Python Script to perform CRUD operations
Run a Python Program to create a DynamoDB table in your AWS account with secondary indexes. 

Skills and Tools: AWS DynamoDB, NoSQL, Python, Boto3, AWS EC2